Reformated how props were being passed so it all froms the store within each component.
