+++ src/App.js
@@ -1,46 +1,23 @@
 import React from "react";
-import { connect } from "react-redux";
 
 import Header from "./components/Header";
 import AddedFeatures from "./components/AddedFeatures";
 import AdditionalFeatures from "./components/AdditionalFeatures";
 import Total from "./components/Total";
-import { buyItem, removeFeature } from "./actions";
 
 const App = props => {
-  const removeFeature = item => {
-    // dispatch an action here to remove an item
-    return props.removeFeature(item);
-  };
-
-  const buyItem = item => {
-    // dipsatch an action here to add an item
-    return props.buyItem(item);
-  };
-
   return (
     <div className="boxes">
       <div className="box">
-        <Header car={props.car} additionalPrice={props.additionalPrice} />
-        <AddedFeatures car={props.car} removeFeature={removeFeature} />
+        <Header />
+        <AddedFeatures />
       </div>
       <div className="box">
-        <AdditionalFeatures
-          additionalFeatures={props.additionalFeatures}
-          buyItem={buyItem}
-        />
-        <Total car={props.car} additionalPrice={props.additionalPrice} />
+        <AdditionalFeatures />
+        <Total />
       </div>
     </div>
   );
 };
 
-const mapStateToProps = state => {
-  return {
-    car: state.car,
-    additionalFeatures: state.additionalFeatures,
-    additionalPrice: state.additionalPrice
-  };
-};
-
-export default connect(mapStateToProps, { buyItem, removeFeature })(App);
+export default App;
