+++ src/components/AddedFeature.js
@@ -1,13 +1,17 @@
 import React from "react";
+import { connect } from "react-redux";
+
+import { removeFeature } from "../actions";
 
 const AddedFeature = props => {
+  const removeFeature = item => {
+    // dispatch an action here to remove an item
+    return props.removeFeature(item);
+  };
   return (
     <li>
       {/* Add an onClick to run a function to remove a feature */}
-      <button
-        onClick={() => props.removeFeature(props.feature)}
-        className="button"
-      >
+      <button onClick={() => removeFeature(props.feature)} className="button">
         X
       </button>
       {props.feature.name}
@@ -15,4 +19,4 @@ const AddedFeature = props => {
   );
 };
 
-export default AddedFeature;
+export default connect(null, { removeFeature })(AddedFeature);
