+++ src/components/AddedFeatures.js
@@ -1,19 +1,17 @@
 import React from "react";
+import { connect } from "react-redux";
 
+import { removeFeature } from "../actions";
 import AddedFeature from "./AddedFeature";
 
 const AddedFeatures = props => {
   return (
     <div className="content">
       <h6>Added features:</h6>
-      {props.car.features.length ? (
+      {props.features.length ? (
         <ol type="1">
-          {props.car.features.map(item => (
-            <AddedFeature
-              key={item.id}
-              feature={item}
-              removeFeature={props.removeFeature}
-            />
+          {props.features.map(item => (
+            <AddedFeature key={item.id} feature={item} />
           ))}
         </ol>
       ) : (
@@ -23,4 +21,10 @@ const AddedFeatures = props => {
   );
 };
 
-export default AddedFeatures;
+const mapStateToProps = state => {
+  return {
+    features: state.car.features
+  };
+};
+
+export default connect(mapStateToProps, { removeFeature })(AddedFeatures);
