+++ src/components/AdditionalFeatures.js
@@ -1,5 +1,5 @@
 import React from "react";
-
+import { connect } from "react-redux";
 import AdditionalFeature from "./AdditionalFeature";
 
 const AdditionalFeatures = props => {
@@ -9,11 +9,7 @@ const AdditionalFeatures = props => {
       {props.additionalFeatures.length ? (
         <ol type="1">
           {props.additionalFeatures.map(item => (
-            <AdditionalFeature
-              key={item.id}
-              feature={item}
-              buyItem={props.buyItem}
-            />
+            <AdditionalFeature key={item.id} feature={item} />
           ))}
         </ol>
       ) : (
@@ -23,4 +19,10 @@ const AdditionalFeatures = props => {
   );
 };
 
-export default AdditionalFeatures;
+const mapStateToProps = state => {
+  return {
+    additionalFeatures: state.additionalFeatures
+  };
+};
+
+export default connect(mapStateToProps, {})(AdditionalFeatures);
