+++ src/components/Header.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { connect } from "react-redux";
 
 const Header = props => {
   return (
@@ -12,4 +13,11 @@ const Header = props => {
   );
 };
 
-export default Header;
+const mapStateToProps = state => {
+  return {
+    car: state.car,
+    additionalPrice: state.additionalPrice
+  };
+};
+
+export default connect(mapStateToProps, {})(Header);
