+++ src/components/Total.js
@@ -1,4 +1,5 @@
-import React from 'react';
+import React from "react";
+import { connect } from "react-redux";
 
 const Total = props => {
   return (
@@ -8,4 +9,11 @@ const Total = props => {
   );
 };
 
-export default Total;
+const mapStateToProps = state => {
+  return {
+    car: state.car,
+    additionalPrice: state.additionalPrice
+  };
+};
+
+export default connect(mapStateToProps, {})(Total);
